requirements.txt
beautifulsoup4==4.11.1
Flask==2.0.2
python-dotenv==0.21.0
selenium==4.7.2
xlsxwriter==3.0.5
